requirements.txt
numpy
pandas
scikit-surprise
pillow
regex
urllib3
tk